requirements.txt
numpy
mysql-connector-python
sqlalchemy
confluent_kafka
pandas
kafka

